+++ src/components/LoadTestfileButton.js
@@ -1,44 +1,13 @@
 import React, { useCallback, useState } from 'react';
-import {
-  Button,
-  Dropdown,
-  DropdownItem,
-  DropdownMenu,
-  DropdownToggle,
-  Icon,
-  Modal,
-  ModalBody,
-  ModalFooter,
-  ModalHeader,
-} from 'design-react-kit';
+import { Dropdown, DropdownItem, DropdownMenu, DropdownToggle, Icon } from 'design-react-kit';
 import { useSelector, useDispatch } from 'react-redux';
-import { createUseStyles } from 'react-jss';
 import { setDocumentUrl, resetValidationResults, setTemplateDocumentName } from '../redux/actions.js';
 import { isValidationInProgress } from '../redux/selectors.js';
 import { TEMPLATE_DOCUMENT_URL } from '../utils.mjs';
-import useModalView from './useModalView.js';
-import { SwaggerPanel } from './SwaggerPanel.js';
-
-const useStyles = createUseStyles({
-  modalFullScreen: {
-    minWidth: '100% !important',
-    margin: '0 !important',
-    padding: '40px',
-  },
-  description: {
-    verticalAlign: 'top',
-    padding: '1em',
-  },
-});
 
 export const LoadTestfileButton = () => {
   const validationInProgress = useSelector((state) => isValidationInProgress(state));
   const dispatch = useDispatch();
-  const [isModalOpen, closeModal, openModal] = useModalView();
-  const handleConfirmAction = useCallback(() => {
-    closeModal();
-  }, [closeModal]);
-
   const handleOnClick = useCallback(
     (url) => {
       console.log(`Loading ${url}`);
@@ -48,7 +17,7 @@ export const LoadTestfileButton = () => {
     },
     [dispatch]
   );
-  const classes = useStyles();
+
   const [open, toggle] = useState(false);
   return (
     <Dropdown isOpen={open} toggle={() => toggle(!open)}>
@@ -61,31 +30,6 @@ export const LoadTestfileButton = () => {
         </DropdownItem>
         <DropdownItem onClick={() => handleOnClick('errorfile.yaml')}>
           <Icon className="left" icon="it-upload" aria-hidden size="sm" /> Example API with error
-        </DropdownItem>
-
-        <DropdownItem onClick={openModal}>
-          <Icon className="left" icon="it-upload" aria-hidden size="sm" />
-          Show schema
-          <Modal
-            className={classes.modalFullScreen}
-            fade={false}
-            isOpen={isModalOpen}
-            role="dialog"
-            centered
-            toggle={closeModal}
-          >
-            <ModalHeader charCode={215} closeAriaLabel="Close" tag="h5" wrapTag="div" toggle={closeModal}>
-              Mostra schema
-            </ModalHeader>
-            <ModalBody className="mt-3" tag="div">
-              <SwaggerPanel />
-            </ModalBody>
-            <ModalFooter tag="div">
-              <Button color="primary" icon={false} onClick={handleConfirmAction} tag="button">
-                Close
-              </Button>
-            </ModalFooter>
-          </Modal>
         </DropdownItem>
       </DropdownMenu>
     </Dropdown>
